+++ lib/ratelimit/constructor_test.go
@@ -21,12 +21,10 @@
 package ratelimit
 
 import (
-	"strings"
 	"testing"
 
 	"github.com/Jeffail/benthos/lib/log"
 	"github.com/Jeffail/benthos/lib/metrics"
-	"github.com/Jeffail/benthos/lib/types"
 	yaml "gopkg.in/yaml.v3"
 )
 
@@ -42,32 +40,6 @@ func TestConstructorBadType(t *testing.T) {
 
 	if _, err := New(conf, nil, log.Noop(), metrics.Noop()); err == nil {
 		t.Error("Expected error, received nil for invalid type")
-	}
-}
-
-func TestConstructorBlockType(t *testing.T) {
-	Constructors["footype"] = TypeSpec{
-		constructor: func(
-			conf Config,
-			mgr types.Manager,
-			log log.Modular,
-			stats metrics.Type,
-		) (types.RateLimit, error) {
-			return nil, nil
-		},
-	}
-
-	conf := NewConfig()
-	conf.Type = "footype"
-
-	Block("footype", "because testing")
-
-	_, err := New(conf, nil, log.Noop(), metrics.Noop())
-	if err == nil {
-		t.Fatal("Expected error, received nil for blocked type")
-	}
-	if !strings.Contains(err.Error(), "because testing") {
-		t.Errorf("Unexpected error: %v", err)
 	}
 }